Fix compile, linting issues
This corrects the linter error by specifying that this will be used to update the existing library, and also corrects the newly-reintroduced compiler error.

This compiler error is because the constant HT16K33_DIM_1 is equal to 0, and since SetBrightness only accepts uint8_t as a data type it is not possible for the passed value to be less than 0. As such specifying `level >= HT16K33_DIM_1` is unnecessary and the cause of a compiler warning in the Arduino IDE.